fix: address more pr issues, improve documentation
