+++ immuni_exposure_ingestion/tasks/delete_old_data.py
@@ -18,7 +18,7 @@ from immuni_common.models.mongoengine.batch_file import BatchFile
 from immuni_exposure_ingestion.celery import celery_app
 from immuni_exposure_ingestion.core import config
 from immuni_exposure_ingestion.models.upload import Upload
-from immuni_exposure_ingestion.monitoring.celery import BATCHES_DELETED, UPLOADS_DELETED
+from immuni_exposure_ingestion.monitoring.celery import BATCH_FILES_DELETED, UPLOADS_DELETED
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -53,6 +53,5 @@ def delete_old_data() -> None:
         extra=dict(n_deleted=batches_deleted, created_before=reference_date),
     )
 
-    # Increase metrics as well
     UPLOADS_DELETED.inc(uploads_deleted)
-    BATCHES_DELETED.inc(batches_deleted)
+    BATCH_FILES_DELETED.inc(batches_deleted)
